+++ src/shared/components/palette/case-link/components/linked-cases-table/linked-cases-to-table.component.spec.ts
@@ -1,7 +1,11 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ActivatedRoute, Router } from '@angular/router';
-import { async, ComponentFixture, getTestBed, TestBed } from '@angular/core/testing';
+import {
+  async,
+  ComponentFixture,
+  TestBed,
+} from '@angular/core/testing';
 import { of, throwError } from 'rxjs';
 import { PipesModule } from '../../../../../pipes/pipes.module';
 import { SearchService } from '../../../../../services';
@@ -10,10 +14,12 @@ import { LinkedCasesToTableComponent } from './linked-cases-to-table.component';
 
 import createSpyObj = jasmine.createSpyObj;
 import { CaseField } from '../../../../../domain';
-import { CommonDataService, LovRefDataByServiceModel } from '../../../../../services/common-data-service/common-data-service';
+import {
+  CommonDataService,
+} from '../../../../../services/common-data-service/common-data-service';
 import { AbstractAppConfig } from '../../../../../../app.config';
-import { CaseLink } from '../../domain';
 import { LinkedCasesService } from '../../services';
+import { mockCaseLinkingReasonCode, mockCaseLinkResponse, mocklinkedCases, mockSearchByCaseIdsResponse } from '../__mocks__';
 
 describe('LinkCasesToTableComponent', () => {
   let component: LinkedCasesToTableComponent;
@@ -24,138 +30,37 @@ describe('LinkCasesToTableComponent', () => {
   let appConfig: any;
   let linkedCasesService: any;
 
-  const linkedCases: CaseLink[] = [
-    {
-      caseReference: '1682374819203471',
-      reasons: [],
-      createdDateTime: '',
-      caseType: 'SSCS',
-      caseState: 'state',
-      caseService: 'Tribunal',
-      caseName: 'SSCS 2.1'
-    },
-    {
-      caseReference: '1682897456391875',
-      reasons: [],
-      createdDateTime: '',
-      caseType: 'SSCS',
-      caseState: 'state',
-      caseService: 'Tribunal',
-      caseName: 'SSCS 2.1'
-    }
-  ];
-
   linkedCasesService = {
     caseId: '1682374819203471',
-    linkedCases: linkedCases
+    linkedCases: mocklinkedCases,
   };
 
   let mockRouter: any;
   mockRouter = {
     navigate: jasmine.createSpy('navigate'),
-    url: ''
+    url: '',
   };
 
-  let mockCaseLinkResponse = [
-    {
-      'caseReference': '1652112127295261',
-      'modified_date_time': '2022-05-10',
-      'caseType': 'Benefit_SCSS',
-      'reasons': [
-        {
-          'reasonCode': 'Same Party',
-          'OtherDescription': ''
-        }
-      ]
-    },
-    {
-      'caseReference': '1652111610080172',
-      'modified_date_time': '2022-05-10',
-      'caseType': 'Benefit_SCSS',
-      'reasons': [
-        {
-          'reasonCode': 'Case consolidated',
-          'OtherDescription': ''
-        }
-      ]
-    },
-    {
-      'caseReference': '1652111179220086',
-      'modified_date_time': '2022-05-10',
-      'caseType': 'Benefit_SCSS',
-      'reasons': [
-        {
-          'reasonCode': 'Progressed as part of this lead case',
-          'OtherDescription': ''
-        },
-        {
-          'reasonCode': 'Familial',
-          'OtherDescription': ''
-        }
-      ]
-    }
-  ];
-  const linkCaseReasons: LovRefDataByServiceModel = {
-    list_of_values: [
-    {
-      key: 'progressed',
-      value_en: 'Progressed as part of this lead case',
-      value_cy: '',
-      hint_text_en: 'Progressed as part of this lead case',
-      hint_text_cy: '',
-      lov_order: 1,
-      parent_key: null,
-      category_key: 'caseLinkReason',
-      parent_category: '',
-      active_flag: 'Y',
-      child_nodes: null,
-      from: 'exui-default',
-    },
-    {
-      key: 'bail',
-      value_en: 'Bail',
-      value_cy: '',
-      hint_text_en: 'Bail',
-      hint_text_cy: '',
-      lov_order: 2,
-      parent_key: null,
-      category_key: 'caseLinkReason',
-      parent_category: '',
-      active_flag: 'Y',
-      child_nodes: null,
-      from: 'exui-default',
-    },
-    {
-      key: 'other',
-      value_en: 'Other',
-      value_cy: '',
-      hint_text_en: 'Other',
-      hint_text_cy: '',
-      lov_order: 3,
-      parent_key: null,
-      category_key: 'caseLinkReason',
-      parent_category: '',
-      active_flag: 'Y',
-      child_nodes: null,
-      from: 'exui-default',
-    },
-  ]};
-
   beforeEach(async(() => {
-    appConfig = createSpyObj<AbstractAppConfig>('appConfig', ['getRDCommonDataApiUrl']);
+    appConfig = createSpyObj<AbstractAppConfig>('appConfig', [
+      'getRDCommonDataApiUrl',
+    ]);
     commonDataService = createSpyObj('commonDataService', ['getRefData']);
-    casesService = createSpyObj('casesService', ['getCaseViewV2', 'getCaseLinkResponses']);
-    searchService = createSpyObj('searchService', ['searchCases', 'searchCasesByIds']);
+    casesService = createSpyObj('casesService', [
+      'getCaseViewV2',
+      'getCaseLinkResponses',
+    ]);
+    searchService = createSpyObj('searchService', [
+      'searchCases',
+      'searchCasesByIds',
+    ]);
     TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule,
-        PipesModule
-      ],
+      imports: [RouterTestingModule, PipesModule],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
       providers: [
         {
           provide: ActivatedRoute,
-          useValue: {snapshot: {data: {case: {case_id: '123'}}}}
+          useValue: { snapshot: { data: { case: { case_id: '123' } } } },
         },
         { provide: Router, useValue: mockRouter },
         { provide: CasesService, useValue: casesService },
@@ -165,50 +70,33 @@ describe('LinkCasesToTableComponent', () => {
         { provide: LinkedCasesService, useValue: linkedCasesService },
       ],
       declarations: [LinkedCasesToTableComponent],
-    })
-      .compileComponents();
+    }).compileComponents();
   }));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(LinkedCasesToTableComponent);
-    component = fixture.componentInstance
-    spyOn(component, 'searchCasesByCaseIds').and.returnValue(of([{results: [{
-      'case_id': '1652111179220086',
-      'supplementary_data': {
-          'HMCTSServiceId': 'BBA3'
-      },
-      'case_fields': {
-          '[JURISDICTION]': 'SSCS',
-          'dwpDueDate': '2022-06-13',
-          '[LAST_STATE_MODIFIED_DATE]': '2022-05-09T15:46:31.153',
-          'caseReference': '22',
-          '[CREATED_DATE]': '2022-05-09T15:46:19.243',
-          'dateSentToDwp': '2022-05-09',
-          '[CASE_REFERENCE]': '1652111179220086',
-          '[STATE]': 'withDwp',
-          '[ACCESS_GRANTED]': 'STANDARD',
-          '[SECURITY_CLASSIFICATION]': 'PUBLIC',
-          '[ACCESS_PROCESS]': 'NONE',
-          '[CASE_TYPE]': 'Benefit_SCSS',
-          'appeal.appellant.name.lastName': 'Torres',
-          'region': 'Quo nostrum vitae re',
-          '[LAST_MODIFIED_DATE]': '2022-05-09T15:46:31.153'
-      }
-  }] }]));
-    component.caseField = <CaseField>({
+    component = fixture.componentInstance;
+    spyOn(component, 'searchCasesByCaseIds').and.returnValue(
+      of([
+        {
+          results: mockSearchByCaseIdsResponse,
+        },
+      ])
+    );
+    component.caseField = <CaseField>{
       id: 'caseLink',
       field_type: {
-      id: 'Text',
-      type: 'Text',
-      complex_fields: []
+        id: 'Text',
+        type: 'Text',
+        complex_fields: [],
       },
       display_context: 'OPTIONAL',
       label: 'First name',
       show_summary_content_option: {},
-      retain_hidden_value:  false,
+      retain_hidden_value: false,
       hidden: false,
-      });
-    commonDataService.getRefData.and.returnValue(of(linkCaseReasons));
+    };
+    commonDataService.getRefData.and.returnValue(of(mockCaseLinkingReasonCode));
     component.caseField.value = mockCaseLinkResponse;
   });
 
@@ -230,7 +118,7 @@ describe('LinkCasesToTableComponent', () => {
     let caseNameMissingEle = 0;
     component.ngOnInit();
     fixture.detectChanges();
-    document.querySelectorAll('a').forEach(item => {
+    document.querySelectorAll('a').forEach((item) => {
       if (item.textContent.includes('Case name missing')) {
         caseNameMissingEle += 1;
       }
@@ -245,7 +133,9 @@ describe('LinkCasesToTableComponent', () => {
   });
 
   it('should render the failure panel when api returns non 200', () => {
-    component.searchCasesByCaseIds = jasmine.createSpy().and.returnValue(throwError({}));
+    component.searchCasesByCaseIds = jasmine
+      .createSpy()
+      .and.returnValue(throwError({}));
     component.ngOnInit();
     spyOn(component.notifyAPIFailure, 'emit');
     fixture.detectChanges();
@@ -258,6 +148,8 @@ describe('LinkCasesToTableComponent', () => {
     component.linkedCasesFromResponse = [];
     component.ngOnInit();
     fixture.detectChanges();
-    expect(document.getElementsByClassName('govuk-table__cell')[0].textContent.trim()).toEqual('None');
+    expect(
+      document.getElementsByClassName('govuk-table__cell')[0].textContent.trim()
+    ).toEqual('None');
   });
 });