+++ src/shared/components/palette/case-link/components/read-linked-cases.component.spec.ts
@@ -1,7 +1,7 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { RouterTestingModule } from '@angular/router/testing';
-import { ActivatedRoute, Router } from '@angular/router';
-import { async, ComponentFixture, getTestBed, TestBed } from '@angular/core/testing';
+import { Router } from '@angular/router';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import createSpyObj = jasmine.createSpyObj;
 import { ReadLinkedCasesComponent } from './read-linked-cases.component';