+++ src/shared/components/palette/case-link/write-case-link-field.component.ts
@@ -78,6 +78,7 @@ export class WriteCaseLinkFieldComponent extends AbstractFieldWriteComponent imp
   }
 
   public hasCaseLinkCollection(): boolean {
-    return this.caseField.field_type && this.caseField.field_type.collection_field_type.id === 'CaseLink';
+    return this.caseField.field_type && this.caseField.field_type.type === 'Collection' &&
+            this.caseField.field_type.collection_field_type.id === 'CaseLink'
   }
 }